+++ qa-core/src/config/internal-api/fixtures/rows.ts
@@ -1,0 +1,8 @@
+import { Row } from "@budibase/types"
+
+export const generateNewRowForTable = (tableId: string): Row => {
+    return {
+        TestColumn: "TestRow",
+        tableId: tableId
+    }
+}
